make int long int

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,14 +9,14 @@
 int main(void)
 {
 	int n = 50, i;
-	int a = 0, b = 1, c;
+	long int a = 0, b = 1, c;
 
-	printf("%d, %d", a, b);
+	printf("%ld, %ld", a, b);
 
 	for (i = 3; i <= n; i++)
 	{
 		c = a + b;
-		printf(", %d", c);
+		printf(", %ld", c);
 		a = b;
 		b = c;
 	}
